refactor : make test utils folder
